Move types from reader into new contract file
