+++ extension/src/plots/index.ts
@@ -13,7 +13,8 @@ import { Experiments } from '../experiments'
 import { Resource } from '../resourceLocator'
 import { InternalCommands } from '../commands/internal'
 import { definedAndNonEmpty } from '../util/array'
-import { ExperimentsOutput, TEMP_PLOTS_DIR } from '../cli/dvc/reader'
+import { ExperimentsOutput } from '../cli/dvc/contract'
+import { TEMP_PLOTS_DIR } from '../cli/dvc/constants'
 import { removeDir } from '../fileSystem'
 import { Toast } from '../vscode/toast'
 import { pickPaths } from '../path/selection/quickPick'
