+++ extension/src/plots/paths/model.ts
@@ -6,7 +6,7 @@ import {
   PlotPath,
   TemplateOrder
 } from './collect'
-import { PlotsOutput } from '../../cli/dvc/reader'
+import { PlotsOutput } from '../../cli/dvc/contract'
 import { PathSelectionModel } from '../../path/selection/model'
 import { PersistenceKey } from '../../persistence/constants'
 import { performSimpleOrderedUpdate } from '../../util/array'
