+++ extension/src/repository/model/index.ts
@@ -17,7 +17,8 @@ import {
   SCMState,
   SourceControlStatus
 } from '../sourceControlManagement'
-import { DataStatusOutput, DvcError, isDvcError } from '../../cli/dvc/reader'
+import { DataStatusOutput, DvcError } from '../../cli/dvc/contract'
+import { isDvcError } from '../../cli/dvc/reader'
 import { Disposable } from '../../class/dispose'
 import { sameContents } from '../../util/array'
 import { Data } from '../data'
