+++ extension/src/test/fixtures/expShow/deeplyNested.ts
@@ -1,4 +1,4 @@
-import { ExperimentsOutput } from '../../../cli/dvc/reader'
+import { ExperimentsOutput } from '../../../cli/dvc/contract'
 import { timestampColumn } from '../../../experiments/columns/constants'
 import {
   Column,
